requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0